pta: stats: fix compilation incompatible pointer warning
To reproduce the issue:
$ make PLATFORM=imx-mx8mmevk CFG_WITH_STATS=y CFG_TA_STATS=y

core/pta/stats.c: In function ‘get_user_ta_stats’:
core/pta/stats.c:169:37: warning: passing argument 2 of ‘tee_ta_instance_stats’ from incompatible pointer type [-Wincompatible-pointer-types]
  169 |                                     &p[0].memref.size);
      |                                     ^~~~~~~~~~~~~~~~~
      |                                     |
      |                                     size_t * {aka long unsigned int *}
In file included from core/include/kernel/pseudo_ta.h:10,
                 from core/pta/stats.c:8:
core/include/kernel/tee_ta_manager.h:171:56: note: expected ‘uint32_t *’ {aka ‘unsigned int *’} but argument is of type ‘size_t *’ {aka ‘long unsigned int *’}
  171 | TEE_Result tee_ta_instance_stats(void *buff, uint32_t *buff_size);
      |                                              ~~~~~~~~~~^~~~~~~~~
core/pta/stats.c: At top level:
cc1: note: unrecognized command-line option ‘-Wno-c2x-extensions’ may have been intended to silence earlier diagnostics

Fixes: 7509620b8b95 ("GP131: Update TEE_Param")
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>

+++ core/include/kernel/tee_ta_manager.h
@@ -168,7 +168,7 @@ static inline struct tee_ta_ctx *to_ta_ctx(struct ts_ctx *ctx)
 }
 
 #if defined(CFG_TA_STATS)
-TEE_Result tee_ta_instance_stats(void *buff, uint32_t *buff_size);
+TEE_Result tee_ta_instance_stats(void *buff, size_t *buff_size);
 #endif
 
 #endif
